+++ types/src/main/java/com/cobber/fta/plugins/identity/BSN_NL.java
@@ -1,0 +1,128 @@
+/*
+ * Copyright 2017-2022 Tim Segall
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.cobber.fta.plugins.identity;
+
+import com.cobber.fta.AnalysisConfig;
+import com.cobber.fta.AnalyzerContext;
+import com.cobber.fta.Facts;
+import com.cobber.fta.FiniteMap;
+import com.cobber.fta.LogicalTypeInfinite;
+import com.cobber.fta.PluginAnalysis;
+import com.cobber.fta.PluginDefinition;
+import com.cobber.fta.core.FTAPluginException;
+import com.cobber.fta.core.FTAType;
+import com.cobber.fta.core.Utils;
+import com.cobber.fta.token.TokenStreams;
+
+/**
+ * Plugin to detect Dutch BSN's (Burger Service Nummer).
+ */
+public class BSN_NL extends LogicalTypeInfinite {
+	/** The Semantic type for this Plugin. */
+	public static final String SEMANTIC_TYPE = "IDENTITY.BSN_NL";
+
+	private static final int LENGTH = 9;
+	private static final String BACKOUT_REGEXP = ".*";
+	private String regExp = BACKOUT_REGEXP;
+
+	/**
+	 * Construct a plugin to detect French SSN's based on the Plugin Definition.
+	 * @param plugin The definition of this plugin.
+	 */
+	public BSN_NL(final PluginDefinition plugin) {
+		super(plugin);
+	}
+
+	@Override
+	public boolean isCandidate(final String trimmed, final StringBuilder compressed, final int[] charCounts, final int[] lastIndex) {
+		if (trimmed.length() - charCounts[' '] != LENGTH)
+			return false;
+
+		for (int i = 0; i < trimmed.length(); i++) {
+			final char ch = trimmed.charAt(i);
+			if (ch == ' ')
+				continue;
+			if (!Utils.isSimpleNumeric(ch))
+				return false;
+		}
+
+		return true;
+	}
+
+	@Override
+	public boolean initialize(final AnalysisConfig analysisConfig) throws FTAPluginException {
+		super.initialize(analysisConfig);
+
+		return true;
+	}
+
+	@Override
+	public String nextRandom() {
+		String base = Utils.getRandomDigits(random, 8);
+
+		return base + VAT.getCheckDigitNL(base);
+	}
+
+	@Override
+	public String getSemanticType() {
+		return SEMANTIC_TYPE;
+	}
+
+	@Override
+	public String getRegExp() {
+		return regExp;
+	}
+
+	@Override
+	public boolean acceptsBaseType(final FTAType type) {
+		return type == FTAType.STRING || type == FTAType.LONG;
+	}
+
+	@Override
+	public FTAType getBaseType() {
+		return FTAType.STRING;
+	}
+
+	@Override
+	public boolean isValid(final String input, final boolean detectMode, final long count) {
+		final StringBuilder b = new StringBuilder(LENGTH);
+
+		for (int i = 0; i < input.length(); i++) {
+			final char ch = input.charAt(i);
+			if (ch == ' ')
+				continue;
+			if (!Utils.isSimpleNumeric(ch))
+				return false;
+			b.append(ch);
+		}
+
+		if (b.length() != LENGTH)
+			return false;
+
+		return b.charAt(LENGTH - 1) == VAT.getCheckDigitNL(b.substring(0, LENGTH - 1));
+	}
+
+	@Override
+	public PluginAnalysis analyzeSet(final AnalyzerContext context, final long matchCount, final long realSamples, final String currentRegExp, final Facts facts, final FiniteMap cardinality, final FiniteMap outliers, final TokenStreams tokenStreams, final AnalysisConfig analysisConfig) {
+
+		if (cardinality.size() < 20 || (double)matchCount/realSamples < getThreshold()/100.0)
+			return new PluginAnalysis(BACKOUT_REGEXP);
+
+		regExp = tokenStreams.getRegExp(false);
+
+		return PluginAnalysis.OK;
+	}
+}