+++ types/src/test/java/com/cobber/fta/TestPlugins.java
@@ -497,6 +497,36 @@ public class TestPlugins {
 	}
 
 	@Test(groups = { TestGroups.ALL, TestGroups.PLUGINS })
+	public void VAT_NL() throws IOException, FTAException {
+		final TextAnalyzer analysis = new TextAnalyzer("BTW-nummer");
+		analysis.setLocale(Locale.forLanguageTag("nl-NL"));
+
+		final String[] inputs = {
+				"NL000000024B01", "NL813195779B01", "NL814170511B01", "NL815216002B01", "NL815498093B01",
+				"NL000000048B01", "NL000000061B01", "NL000000073B01", "NL000000085B01", "NL000000103B01",
+				"NL000000115B01", "NL000000127B01", "NL000000140B01", "NL000000152B01", "NL000000164B01",
+				"NL000000188B01", "NL000000205B01", "NL000000206B01", "NL000000206B01", "NL001079293B01",
+				"NL001368023B01", "NL003156709B01", "NL004909665B07", "NL005033019B01", "NL006292227B01",
+				"NL010000445B01", "NL010000446B01", "NL121745417B01", "NL128297906B01", "NL147804668B01",
+				"NL173389909B01", "NL208560129B01", "NL800272912B01", "NL805332674B01", "NL805969317B01",
+				"NL806825790B01", "NL806925206B01", "NL809442127B01", "NL810195835B01", "NL810876334B01",
+		};
+
+		for (final String input : inputs)
+			analysis.train(input);
+
+		final TextAnalysisResult result = analysis.getResult();
+
+		assertEquals(result.getSampleCount(), inputs.length);
+		assertEquals(result.getType(), FTAType.STRING);
+		assertEquals(result.getSemanticType(), "IDENTITY.VAT_NL");
+		assertEquals(result.getNullCount(), 0);
+		assertEquals(result.getRegExp(), "(NL)?\\d{9}B\\d{2}");
+		assertEquals(result.getMatchCount(), inputs.length - 2);
+		assertEquals(result.getConfidence(), 0.975);
+	}
+
+	@Test(groups = { TestGroups.ALL, TestGroups.PLUGINS })
 	public void VAT_PL() throws IOException, FTAException {
 		final TextAnalyzer analysis = new TextAnalyzer("NIP");
 		analysis.setLocale(Locale.forLanguageTag("pl-PL"));